CURSOR PROJECTS/requirements.txt
flask==2.3.3
werkzeug==2.3.7
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
python-dotenv==0.19.0
waitress==3.0.0
whitenoise==6.6.0
cachetools==5.3.2
flask-caching==2.1.0
gunicorn==21.2.0 